+++ src/main/resources/TimeGuard.fxml
@@ -14,7 +14,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.control.ListView?>
 <VBox prefHeight="600.0" prefWidth="300.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="pl.agh.cs.io.Controller.TimeGuardController">
+      fx:controller="pl.agh.cs.io.controller.TimeGuardController">
   <children>
     <MenuBar VBox.vgrow="NEVER">
       <menus>
