+++ src/main/resources/statistics.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.cell.PropertyValueFactory?>
 <VBox prefHeight="600.0" prefWidth="450.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="pl.agh.cs.io.Controller.StatsController">
+      fx:controller="pl.agh.cs.io.controller.StatsController">
     <TableView fx:id="data">
         <columns>
             <TableColumn text="Name" fx:id="name">
